python_space/requirements.txt
allure-pytest==2.7.0
allure-python-commons==2.7.0
apipkg==1.5
appdirs==1.4.3
Appium-Python-Client==0.26
atomicwrites==1.1.5
attrs==18.1.0
beautifulsoup4==4.7.0
bs4==0.0.1
certifi==2018.11.29
chardet==3.0.4
Click==7.0
colorama==0.3.9
cssselect==1.0.3
ddt==1.1.2
decorator==4.3.0
DingtalkChatbot==1.3.0
enum34==1.1.6
execnet==1.6.0
fake-useragent==0.1.11
Flask==1.0.2
Flask-JWT-Extended==3.15.0
get==1.0.2
idna==2.6
importlib-metadata==0.18
itsdangerous==1.1.0
Jinja2==2.10
lightpush==0.1.3
lxml==4.3.0
MarkupSafe==1.1.0
more-itertools==4.3.0
namedlist==1.7
nose==1.3.7
ntplib==0.3.3
numpy==1.15.4
packaging==19.0
pandas==0.23.4
parameterized==0.6.1
parse==1.9.0
Pillow==5.4.1
pluggy==0.12.0
post==1.0.1
public==1.0.1
py==1.5.4
pyee==5.0.0
PyJWT==1.7.1
PyMySQL==0.8.0
pyparsing==2.4.0
pyppeteer==0.0.25
pyquery==1.4.0
pytest==4.6.3
pytest-forked==1.0.2
pytest-html==1.19.0
pytest-metadata==1.7.0
pytest-rerunfailures==7.0
pytest-xdist==1.29.0
python-dateutil==2.7.5
pytz==2018.7
PyYAML==3.13
query-string==1.0.1
redis==3.0.1
request==1.0.1
requests==2.18.4
requests-html==0.9.0
retry==0.9.2
selenium==3.11.0
six==1.12.0
soupsieve==1.6.2
tqdm==4.28.1
urllib3==1.22
w3lib==1.19.0
wcwidth==0.1.7
websockets==7.0
Werkzeug==0.14.1
wrapcache==1.0.8
xlrd==1.1.0
xlwt==1.3.0


